requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
asttokens
attrs
certifi
charset-normalizer
click
colorama
comm
contourpy
cycler
datasets
debugpy
decorator
dill
executing
filelock
fonttools
frozenlist
fsspec
huggingface-hub
idna
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2
joblib
jupyter_client
jupyter_core
kiwisolver
MarkupSafe
matplotlib
matplotlib-inline
mpmath
multidict
multiprocess
nest-asyncio
networkx
nltk
numpy
packaging
pandas
parso
pillow
platformdirs
prompt_toolkit
propcache
psutil
pure_eval
pyarrow
Pygments
pyparsing
python-dateutil
pytz
PyYAML
pyzmq
regex
requests
safetensors
scikit-learn
scipy
six
stack-data
sympy
threadpoolctl
tokenizers
torch
tornado
tqdm
traitlets
transformers
typing_extensions
tzdata
urllib3
wcwidth
xxhash
yarl
evaluate
nltk
absl-py
rouge_score
transformers
evaluate